+++ src/frontend/src/lib/components/UI/BottomCardOrModal.svelte
